Moving username to df helpers back
